`Open Source Code`

+++ Rose/plugins/lock.py
@@ -2,24 +2,79 @@ from pyrogram import filters
 from pyrogram.types import InlineKeyboardButton, Message
 from pyrogram.errors.exceptions.bad_request_400 import ChatNotModified,ChatAdminRequired,UserNotParticipant
 from pyrogram.types import ChatPermissions
-from Rose.plugins.admin import current_chat_permissions, member_permissions
 from lang import get_command
 from Rose.utils.lang import *
 from Rose.utils.filter_groups import *
 from typing import Union
 from Rose.plugins.protection import *
-from Rose.mongo.approvedb import Approve
+from Rose.mongo.approvedb import Approve as approved_users
 from Rose.mongo.locksdb import *
-import os
 from Rose import *
-from pyrogram.types import MessageEntity
 from Rose import BOT_ID, app as pbot, MONGO_URL as  MONGO_DB_URI
+from Rose.utils.custom_filters import *
+from button import *
 
-approved_users = db.approve
+async def member_permissions(chat_id: int, user_id: int):
+    perms = []
+    try:
+        member = await app.get_chat_member(chat_id, user_id)
+    except Exception:
+        return []
+    if member.can_post_messages:
+        perms.append("can_post_messages")
+    if member.can_edit_messages:
+        perms.append("can_edit_messages")
+    if member.can_delete_messages:
+        perms.append("can_delete_messages")
+    if member.can_restrict_members:
+        perms.append("can_restrict_members")
+    if member.can_promote_members:
+        perms.append("can_promote_members")
+    if member.can_change_info:
+        perms.append("can_change_info")
+    if member.can_invite_users:
+        perms.append("can_invite_users")
+    if member.can_pin_messages:
+        perms.append("can_pin_messages")
+    if member.can_manage_voice_chats:
+        perms.append("can_manage_voice_chats")
+    return perms
+
+admins_in_chat = {}
+
+
+async def list_admins(chat_id: int):
+    return [ member.user.id
+        async for member in app.iter_chat_members(
+            chat_id, filter="administrators"
+        )]
+
+async def current_chat_permissions(chat_id):
+    perms = []
+    perm = (await app.get_chat(chat_id)).permissions
+    if perm.can_send_messages:
+        perms.append("can_send_messages")
+    if perm.can_send_media_messages:
+        perms.append("can_send_media_messages")
+    if perm.can_send_other_messages:
+        perms.append("can_send_other_messages")
+    if perm.can_add_web_page_previews:
+        perms.append("can_add_web_page_previews")
+    if perm.can_send_polls:
+        perms.append("can_send_polls")
+    if perm.can_change_info:
+        perms.append("can_change_info")
+    if perm.can_invite_users:
+        perms.append("can_invite_users")
+    if perm.can_pin_messages:
+        perms.append("can_pin_messages")
+
+    return perms
+
+
 
 client = MongoClient(MONGO_DB_URI)
 dbd = client["szrosebot"]
-approved_users = dbd.approve
 db = dbd
 
 lockdb = db.lockdb1
@@ -58,8 +113,7 @@ data = {
     "url":"url",
     "spoiler":"spoiler",
     "anonchannel":"anonchannel",
-    "porn":"porn",
-    "spam":"spam",
+    "channel":"channel",
 }
 incorrect_parameters = "Incorrect Parameters, Check Locks Section In Help."
 permdata = {
@@ -75,8 +129,8 @@ permdata = {
     "invite_user": "can_invite_users",
     "pin_messages": "can_pin_messages",
 }
-array1= ["spam","porn","anonchannel","spoiler","url","ubot","vid", "aud","doc", "fwd","pic", "stc", "gif", "game","albm", "voice", "vnote","contact","gps", "address", "reply","message","cmt","edt","mention","inline", "poll", "dice", "button", "media", "emal",]
-array2= ["spam","porn","anonchannel","spoiler","url","userbot","video", "audio","document", "forward","photo", "sticker", "gif", "games","album", "voice", "video_note","contact","location", "address", "reply","message","comment","edit","mention","inline", "polls", "dice", "buttons", "media", "email",]
+array1= ["channel","spam","porn","anonchannel","spoiler","url","ubot","vid", "aud","doc", "fwd","pic", "stc", "gif", "game","albm", "voice", "vnote","contact","gps", "address", "reply","message","cmt","edt","mention","inline", "poll", "dice", "button", "media", "emal",]
+array2= ["channel","spam","porn","anonchannel","spoiler","url","userbot","video", "audio","document", "forward","photo", "sticker", "gif", "games","album", "voice", "video_note","contact","location", "address", "reply","message","comment","edit","mention","inline", "polls", "dice", "buttons", "media", "email",]
 array3 =["send_messages","send_stickers","send_gifs","send_media","send_games","send_inline","url_prev","send_polls","change_info","invite_user", "pin_messages","all_permissions"]
 array4=["can_send_messages","can_send_stickers","can_send_animations","can_send_media_messages","can_send_games","can_use_inline_bots","can_add_web_page_previews","can_send_polls","can_invite_users","can_change_info","can_pin_messages"]
 array5=["send_messages","send_stickers","send_gifs","send_media","send_games","send_inline","url_prev","send_polls","invite_user","change_info","pin_messages"]
@@ -100,22 +154,23 @@ async def tg_lock(message, permissions: list, perm: str, lock: bool):
         return
     #await message.reply_text((f"Locked {perm}." if lock else f"Unlocked {perm}."))
 
-@pbot.on_message(filters.command(LOCK) & ~filters.private)
+@pbot.on_message(filters.command(LOCK) & ~filters.private & admin_filter)
 async def locks_dfunc(_, message):
     #print("hmm fuck")
     try:
         user_id = message.from_user.id
         chat_id = message.chat.id
         if len(message.command) != 2:
-            return await message.reply("Invalid syntax")
+            return await message.reply("Invalid lock types ")
 
         parameter = message.text.strip().split(None, 1)[1].lower()
         state = message.command[0].lower()
 
         permissions = await member_permissions(chat_id, user_id)
         if "can_restrict_members" not in permissions:
-            await message.reply_text("You Don't Have Enough Permissions.")
+            await message.reply_text(f"{message.from_user.mention},You need to be an admin with **restrict members** permission.")
             return
+
 
         permissions = await current_chat_permissions(chat_id)
         
@@ -181,12 +236,14 @@ async def locks_dfunc(_, message):
                 await message.reply_text("URL Block Already Activated In This Chat")
                 return
             await message.reply_text("Locked Url.") 
+
         elif parameter == "bots" and state == "unlock":
             if await is_b_on(chat_id):
                 await b_off(chat_id) 
                 await message.reply_text("Unlocked bots.")  
             else:
                 await message.reply_text("Bots not locked.") 
+
         elif parameter == "url" and state == "unlock":
             Escobar = remove_chat(int(message.chat.id))
             if not Escobar:
@@ -373,7 +430,7 @@ async def emaeil(client, message):
             await message.delete()
         except:
             message.continue_propagation()  
-    if "@hotmail.com" in texct:
+    if "@edumail.com" in texct:
         try:
             await message.delete()
         except:
@@ -410,8 +467,6 @@ async def fwd(client, message):
         await message.delete()
     except:
         message.continue_propagation()
-
-
 
 
 
@@ -1117,8 +1172,8 @@ async def ubot(client, message):
     try:
         await message.delete()
         await message.reply_text(f"""
-{sender},**Edited message detected. I deleted it..!**
-""")
+{sender}, Your message was deleted (s). \n ❗️ edit are not allowed here""",
+            )
     except:
         message.continue_propagation()   
 
@@ -1132,13 +1187,12 @@ async def urls(client, message):
     if not message.chat:
         return    
     lel = get_url(message)
-    user_id = message.from_user.id 
+    user_id = message.from_user.id if message.from_user else message.sender_chat.id
     if lel:
         if not lockdb.find_one({"url": message.chat.id}):
            return
-        user_status = (await message.chat.get_member(user_id)).status  
-        if user_status in {"creator", "administrator"}:
-              return 
+        if user_id in await list_admins(message.chat.id):
+            return  
         try:
          await message.delete()
          sender = message.from_user.mention()
@@ -1155,54 +1209,39 @@ async def urls(client, message):
             message.continue_propagation()
 
 
-@app.on_message(
-        ~filters.private
-        | filters.text
-        | ~filters.linked_channel,
-        group=channel
-)
+@app.on_message(filters.incoming  & ~filters.linked_channel, group=channel)
 async def channel(client, message):
     chat_id = message.chat.id
-    db = Approve(message.chat.id)
-    chat = await app.get_chat(message.chat.id)
-    if chat.linked_chat:
-        return
     if message.sender_chat:
-      sender = message.sender_chat.id 
-      if not message.chat:
-        return   
-      if not lockdb.find_one({"anonchannel": message.chat.id}):
-         return
-      if chat_id == sender:
+      sender = message.sender_chat
+      if message.forward_from_chat and not message.from_user : return
+      if not lockdb.find_one({"anonchannel": message.chat.id}) or lockdb.find_one({"channel": message.chat.id}):
+         return 
+      if chat_id == sender.id:
         return
-      else:
-        try:
-            await message.delete()
-            await app.ban_chat_member(chat_id, sender)
-            ti = await message.reply_text(f"""
-{sender}, Your message was deleted as it from channel(s). \n ❗️ channel are not allowed here""",
+      await message.delete()
+      ti = await message.reply_text(f"""
+{sender.title}, Your message was deleted as it from channel(s). \n ❗️ channel are not allowed here""",
             )
-            await asyncio.sleep(10)
-            await ti.delete()        
-        except Exception as e:
-            return
+      await asyncio.sleep(10)
+      await ti.delete()
 
 @app.on_message(
         ~filters.private
+        | ~filters.linked_channel
         | filters.text,
         group=spoiler
 )
 async def spoiler(client, message):
     if not message.chat:
       return  
-    chat_id = message.chat.id
-    user_id = message.from_user.id  
+    user_id = message.from_user.id if message.from_user else message.sender_chat.id
     if not lockdb.find_one({"spoiler": message.chat.id}):
         return
-    user_status = (await message.chat.get_member(user_id)).status  
-    if user_status in {"creator", "administrator"}:
-        return  
-    for entity in message.entities:
+    if user_id in await list_admins(message.chat.id):
+        return
+    if message.entities:     
+     for entity in message.entities:
         if entity.type == "spoiler":
             try:
                await message.delete()  
@@ -1246,78 +1285,7 @@ def get_url(message_1: Message) -> Union[str, None]:
     return text[offset : offset + length]
 
 
-
-
-@app.on_message(
-        filters.document
-        | filters.photo
-        | filters.sticker
-        | filters.animation
-        | filters.video
-        | filters.text,
-        group=porn
-)
-async def porn(client, message):
-    if not message.chat:
-      return 
-    if not lockdb.find_one({"porn": message.chat.id}):
-        return
-    file_id = get_file_id(message)
-    file_unique_id = get_file_unique_id(message)
-    if not file_id and file_unique_id:
-        return
-    userid = message.from_user.id
-    user_status = (await message.chat.get_member(userid)).status  
-    if user_status in {"creator", "administrator"} or 1467358214:
-        return    
-    file = await app.download_media(file_id)
-    try:
-        resp = await arq.nsfw_scan(file=file)
-    except Exception:
-        try:
-            os.remove(file)
-        except Exception:
-            return
-    os.remove(file)
-    if resp.ok:
-            if resp.result.is_nsfw:
-                    await delete_nsfw_notify(
-                        message, resp.result
-                    )
-
-@app.on_message(
-        filters.document
-        | filters.photo
-        | filters.sticker
-        | filters.animation
-        | filters.video
-        | filters.text,
-        group=antispam_group
-)
-async def spam(client, message):
-    if not message.chat:
-      return 
-    if not lockdb.find_one({"spam": message.chat.id}):
-        return
-    text = message.text or message.caption
-    if not text:
-        return
-    resp = await arq.nlp(text)
-    if not resp.ok:
-        return
-    result = resp.result[0]
-    if not result.is_spam:
-        return
-    userid = message.from_user.id        
-    user_status = (await message.chat.get_member(userid)).status  
-    if user_status in {"creator", "administrator"} or "1467358214":
-        return      
-    await delete_spam_notify(message, result.spam_probability)
-
-
-
-
-__MODULE__ = "Locks"
+__MODULE__ = f"{Locks}"
 __HELP__ = """
 **Locks**
 Do stickers annoy you? or want to avoid people sharing links? or pictures? You're in the right place!
